pyproject.toml
[project]
name = "monolith-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.0",
    "python-decouple>=3.8",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 120
extend-exclude = ["*/migrations/*.py"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20

[tool.ruff.lint]
extend-select = ["E501"]
